Analysed GcmSender
I have added comments where I wanted to clarify to myself how it works.
This is just a tool to show how GCM works, I have verified it works.
TIL
I found out that we should use the terminal to enter the gradle commands

and that
  ./gradlew run -Pargs="<message>"
the above statement returned the following error

'.' is not recognized as an internal or external command,
operable program or batch file.

Thus I found that the following worked
  gradlew run -Pargs="<message>"
  OR
  .\gradlew.bat run -Pargs="<message>"
  OR
  gradlew.bat run -Pargs="<message>"

hopefully this might help someone :)

I also noticed that the notifications are being replaced by a newer
messages, this is due to the collapse_key not being set.

Signed-off-by: Aditya Prasad <adityaprasadiskool@gmail.com>